Implement an state machine
Add an interface to add checkpoints in each stage. Each function represents an stage, which is any job (compile the loader, the eBPF program, deploy of the compiled eBPF object, etc), each stage has a JobTransition which saves the current state. For now, it works only with a simple log but in a next iteration I'll add a persistent storage.